requirements.txt
matplotlib==3.3.3
pandas==1.3.2
python-dotenv==0.19.2
requests==2.23.0
SQLAlchemy==1.4.15
python-dotenv==0.13.0
lxml==4.6.3